Adds FINOS Legend Bundle Integration tests
Adds the linting and integration test github workflows which will be triggered
when a PR is created, when it merges, and when a manual workflow is dispatched.

The integration tests expect the GITLAB_CLIENT_ID and GITLAB_CLIENT_SECRET env
variables to be set.

The integration tests will deploy the locally defined bundle.yaml on microk8s,
configures the necessary gitlab-client-id and gitlab-client-secret config options
for gitlab-integrator. The integration tests will expect that the deployed units
become active, that FINOS Legend Engine, SDLC, and Studio are reachable through
their IPs and through the default legend-host external-hostname configured. The
Legend applications should be reachable through a newly configured external-hostname.