Adds publishing jobs
Adds the publish_images.yaml workflow, which will run everyday or when manually dispatched.

If there is a new release that was not yet published on charmhub, it will upload
the respective FINOS Legend Engine, SDLC, and Studio images into Charmhub and
make a release. It will also create a new release on latest. This workflow will then
trigger create_release.yaml.

Adds the create_release.yaml workflow, which can also be manually triggered.

This workflow will create a new branch with the given release name.

The following Secrets need to be added (Settings > Actions > New repository secret):

- CHARMCRAFT_AUTH: needed to upload the FINOS Legend images, upload the bundle,
  release the charms and bundle.
- GH_TOKEN: needed to chain the 2 workflows mentioned above.