integration: Removes GITLAB_CLIENT_ID and GITLAB_CLIENT_SECRET env vars
The integration tests only test that the Legend applications are reachable,
which doesn't require valid GITLAB_CLIENT_ID and GITLAB_CLIENT_SECRET configurations.

This is needed because we won't be able to get the GITLAB_CLIENT_ID and
GITLAB_CLIENT_SECRET repository secrets on pull_request (they are fetched from the
repository that makes the pull request), which means that it would fail.