Rename secret CHARMCRAFT_AUTH -> CHARMHUB_TOKEN
The bundle uses `CHARMCRAFT_AUTH` for the secret name, but all the other
juju repos use `CHARMHUB_TOKEN`. This should eliminate the confusion
between the 2 names since they have the same value.

Please note that inside the `create_release.yaml` and
`publish_images.yaml` the `CHARMCRAFT_AUTH` is still used as the
environment variable since that is the name expected by `charmcraft`.